+++ src/terminal7.js
@@ -255,30 +255,10 @@ export class Terminal7 {
         })
 
         // settings button and modal
-        var modal   = document.getElementById("settings-modal")
-        modal.addEventListener('click',
-            async () => {
-                document.getElementById("dotfile-button").classList.remove("on")
-                await this.clear()
-            }
-        )
         document.getElementById("dotfile-button")
-                .addEventListener("click", ev => this.toggleSettings(ev))
-        modal.querySelector(".close").addEventListener('click',
-            async () => {
-                document.getElementById("dotfile-button").classList.remove("on")
-                await this.clear()
-            }
-        )
-        modal.querySelector(".save").addEventListener('click',
-            () => this.wqConf())
-        modal.querySelector(".copy").addEventListener('click',
-            async () => {
-                var area = document.getElementById("edit-conf")
-                this.confEditor.save()
-                Clipboard.write({string: area.value})
-                await this.clear()
-            })
+                .addEventListener("click", () => this.map.shell.confEditor ?
+                    this.map.shell.closeConfig(false)
+                    : this.map.shell.runCommand("config", []))
         this.map.open().then(() => {
            this.goHome()
            setTimeout(async () => {
@@ -1056,6 +1036,24 @@ export class Terminal7 {
         this.keys = {publicKey: publicKey, privateKey: privateKey}
         this.lastIdVerify = now
         return this.keys
+    }
+    async getDotfile() {
+        return (await Preferences.get({key: "dotfile"})).value || DEFAULT_DOTFILE
+    }
+    saveDotfile() {
+        this.cells.forEach(c => {
+            if (typeof(c.setTheme) == "function")
+                c.setTheme(this.conf.theme)
+        })
+        if (this.pb &&
+            ((this.pb.host != this.conf.net.peerbook) 
+             || (this.pb.peerName != this.conf.peerbook.peer_name)
+             || (this.pb.insecure != this.conf.peerbook.insecure)
+             || (this.pb.email != this.conf.peerbook.email))) {
+            this.pbClose()
+            this.pbConnect()
+        }
+        return Preferences.set({key: "dotfile", value: TOML.stringify(this.conf)})
     }
 }