+++ CHANGELOG.md
@@ -6,13 +6,15 @@ will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unrelease
+## Unreleased
 
 ### Fixed
 
 - added a reset for out-of-sync peers
 - fixing reconnect failures
 - fixing peerbook disconnections
+- dotfile editor now open inside TWR
+- Terminal7 is now T7 on narrow screens
 
 ## 1.4.2 - 2023/1/31